ADD IT

+++ core/src/main/java/org/opensearch/sql/expression/function/udf/SpanFunction.java
@@ -54,10 +54,11 @@ public class SpanFunction extends ImplementorUDF {
             OperandTypes.family(SqlTypeFamily.STRING, SqlTypeFamily.NUMERIC, SqlTypeFamily.STRING)
                 .or(
                     OperandTypes.family(
-                        SqlTypeFamily.DATETIME, SqlTypeFamily.INTEGER, SqlTypeFamily.STRING))
+                        SqlTypeFamily.DATETIME, SqlTypeFamily.NUMERIC, SqlTypeFamily.STRING))
+                // TODO: numeric span should support decimal as its interval
                 .or(
                     OperandTypes.family(
-                        SqlTypeFamily.INTEGER, SqlTypeFamily.INTEGER, SqlTypeFamily.ANY)));
+                        SqlTypeFamily.NUMERIC, SqlTypeFamily.INTEGER, SqlTypeFamily.ANY)));
   }
 
   public static class SpanImplementor implements NotNullImplementor {
